Update _index.md
Comes-from: Kerry Knopp

```Since I come in here with problems, I wanted to also come in and report a very successful weekend of upgrades. We got 4 sites from as old as <8.2.73 all to 10.3.1 with largely just sticking to important releases along the way.  We have a few that still have issues, but it seems the recipe for success is:
8.2.7.3 -> 9.2.11.3 -> 9.2.13.0 -> 10.3.1
thanks for all the work along the way and for being helpful.```